api/requirements.txt
fastapi>=0.115.12
uvicorn>=0.34.2
sqlmodel>=0.0.24
pydantic>=2.11.4
python-dotenv>=1.1.0